--- src/main/java/AlreadyUnmarkedException.java
@@ -1,6 +1,0 @@
-class RedundantStateChangeException extends RuntimeException{
-    public RedundantStateChangeException(String msg) {
-        super(msg);
-    }
-
-}
