+++ src/main/java/Commands/Command.java
@@ -1,3 +1,5 @@
+package Commands;
+
 public interface Command {
     void execute(Object input);
 }
