+++ src/main/java/Commands/EndCommand.java
@@ -1,0 +1,8 @@
+package Commands;
+
+public class EndCommand implements Command {
+
+    @Override
+    public void execute(Object o) {
+    }
+}
