+++ src/main/java/CustomExceptions/RedundantStateChangeException.java
@@ -1,0 +1,8 @@
+package CustomExceptions;
+
+public class RedundantStateChangeException extends RuntimeException{
+    public RedundantStateChangeException(String msg) {
+        super(msg);
+    }
+
+}
