+++ src/main/java/Tasks/ToDoTask.java
@@ -1,4 +1,6 @@
-class ToDoTask extends Task{
+package Tasks;
+
+public class ToDoTask extends Task {
     public ToDoTask(String description) {
         super(description, TaskType.T);
     }
